fix: 탈퇴 후 재가입 DB 매핑 중복 방지 수정 (#146) (#149)

+++ refit-backend/src/main/java/org/refit/refitbackend/domain/user/repository/UserJobRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UserJobRepository extends JpaRepository<UserJob, Integer> {
+ public interface UserJobRepository extends JpaRepository<UserJob, Long> {
     void deleteByUser_Id(Long userId);
 
     @Query("""
